renamed resources file
