+++ src/resources.rs
@@ -1,3 +1,3 @@
-pub(crate) mod resources;
+pub(crate) mod resource;
 pub(crate) mod retrieve;
 pub(crate) mod asset;
